pyproject.toml
[project]
name = "quokka-agent"
version = "0.1.0"
description = "Local lightweight agent for PC automation with remote control via Telegram"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Quokka Agent Team"}
]
keywords = ["agent", "automation", "telegram", "ollama", "local-llm"]

dependencies = [
    "python-telegram-bot>=21.0",
    "ollama>=0.1.0",
    "openai>=1.0.0",
    "httpx>=0.25.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
    "pdfplumber>=0.10.0",
    "python-magic>=0.4.27",
    "aiofiles>=23.0.0",
    "rich>=13.0.0",
    "tenacity>=8.0.0",
    "watchdog>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
quokka = "quokka.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["quokka"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true
